refactor(server): remove legacy HTTP infrastructure after Elysia migration
Delete all legacy route files (20 .ts files replaced by .elysia.ts
equivalents) and supporting infrastructure (context, routeTable,
routeGuards, guards, json). Extract elysiaJson + API_SECURITY_HEADERS
from lib/elysia-adapter.ts into lib/apiResponse.ts. Remove unused
RouteContext re-export and stale checkGuards comment.

- Delete 7 legacy lib files + routeTable.test.ts
- Delete 20 legacy route files
- New: lib/apiResponse.ts (elysiaJson + API_SECURITY_HEADERS)
- Update all imports (elysia-app.ts, .elysia.ts plugins, lib helpers)
- Remove _ctx: RouteContext param from notifications.ts
- Update docs/elysia-migration.md and oxlint.config.ts
