--- packages/server/src/lib/json.ts
@@ -1,8 +1,0 @@
-import { SECURITY_HEADERS } from './securityHeaders';
-
-export function json(data: unknown, status = 200): Response {
-  return new Response(JSON.stringify(data), {
-    status,
-    headers: { 'Content-Type': 'application/json', ...SECURITY_HEADERS },
-  });
-}
