+++ packages/server/src/routes/filters.elysia.ts
@@ -2,7 +2,7 @@ import { eq } from 'drizzle-orm';
 import { type Elysia } from 'elysia';
 /* eslint-disable @typescript-eslint/no-unsafe-type-assertion */
 
-import { elysiaJson as json } from '../lib/elysia-adapter';
+import { elysiaJson as json } from '../lib/apiResponse';
 import { requireAdminOrPermission } from '../lib/elysia-guards';
 import { EQ_BAND_COLUMNS, eqBandsFromRow } from '../lib/eqBands';
 import { db, tables } from '../shared/db';
