+++ packages/server/src/routes/lowPass.elysia.ts
@@ -3,7 +3,7 @@ import { type Elysia } from 'elysia';
 import * as v from 'valibot';
 /* eslint-disable @typescript-eslint/no-unsafe-type-assertion */
 
-import { elysiaJson as json } from '../lib/elysia-adapter';
+import { elysiaJson as json } from '../lib/apiResponse';
 import { requireAdminOrPermission } from '../lib/elysia-guards';
 import { syncAllFilters } from '../lib/syncAllFilters';
 import { db, tables } from '../shared/db';
