+++ packages/server/src/routes/tags.elysia.ts
@@ -3,7 +3,7 @@ import { type Elysia } from 'elysia';
 import * as v from 'valibot';
 /* eslint-disable @typescript-eslint/no-unsafe-type-assertion */
 
-import { elysiaJson as json } from '../lib/elysia-adapter';
+import { elysiaJson as json } from '../lib/apiResponse';
 import { requireAdminOrPermission, requireAuth } from '../lib/elysia-guards';
 import { emitPlaylistUpdated } from '../lib/socket';
 import { db, tables } from '../shared/db';
